Rename mocks module to fixtures
